Drop Python 2 support
Remove Python 2 backwards compatibility workarounds and mark the package
as Python 3 only.

Python 3.4 was marked as the lowest supported version per package
classifiers, though whether this is still the case hasn't been checked;
the CI workflow for the original repo hasn't run tests on Python 3.4
since December 2020. The package is now marked as compatible with 3.6
and up.